Consolidate functionality in NestedHostColumnVector and HostColumnVector
