+++ java/src/main/java/ai/rapids/cudf/HostColumnVector.java
@@ -18,9 +18,6 @@
 
 package ai.rapids.cudf;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.List;
@@ -33,177 +30,16 @@ import java.util.function.Consumer;
  * to release it.  Call close to decrement the reference count when you are done with the column,
  * and call incRefCount to increment the reference count.
  */
-public final class HostColumnVector implements AutoCloseable {
+public final class HostColumnVector extends BaseHostColumnVector {
   /**
    * The size in bytes of an offset entry
    */
   static final int OFFSET_SIZE = DType.INT32.sizeInBytes;
-  private static final Logger log = LoggerFactory.getLogger(HostColumnVector.class);
-
   static {
     NativeDepsLoader.loadNativeDeps();
   }
 
-  protected static class NestedHostColumnVector implements AutoCloseable {
-
-    private HostMemoryBuffer data;
-    private HostMemoryBuffer offsets;
-    private HostMemoryBuffer validity;
-    private DType type;
-    private long rows;
-    private Optional<Long> nullCount = Optional.empty();
-    private List<NestedHostColumnVector> nestedChildren = new ArrayList<>();
-
-    public NestedHostColumnVector(DType type, long rows,
-                                  Optional<Long> nullCount, HostMemoryBuffer data, HostMemoryBuffer validity,
-                                  HostMemoryBuffer offsets, List<NestedHostColumnVector> nestedChildren) {
-      this.data = data;
-      this.offsets = offsets;
-      this.validity = validity;
-      this.type = type;
-      this.rows = rows;
-      this.nullCount = nullCount;
-      this.nestedChildren = nestedChildren;
-    }
-
-    @Override
-    public void close() {
-      for (NestedHostColumnVector child : nestedChildren) {
-        if (child != null) {
-          child.close();
-        }
-      }
-      if (data != null) {
-        data.close();
-      }
-      if (validity != null) {
-        validity.close();
-      }
-      if (offsets != null) {
-        offsets.close();
-      }
-    }
-
-    HostMemoryBuffer getData() {
-      return data;
-    }
-
-    HostMemoryBuffer getValidity() {
-      return validity;
-    }
-
-    DType getType() {
-      return type;
-    }
-
-    Optional<Long> getNullCount() {
-      return nullCount;
-    }
-
-    List<NestedHostColumnVector> getNestedChildren() {
-      return nestedChildren;
-    }
-
-    public HostMemoryBuffer getOffsets() {
-      return offsets;
-    }
-
-    public long getRows() {
-      return rows;
-    }
-
-
-    Object getElement(int rowIndex) {
-      if (type == DType.LIST) {
-        List retList = new ArrayList();
-        int start = offsets.getInt(rowIndex * DType.INT32.getSizeInBytes());
-        int end = offsets.getInt((rowIndex + 1) * DType.INT32.getSizeInBytes());
-        for (int j = start; j < end; j++) {
-          retList.add(nestedChildren.get(0).getElement(j));
-        }
-        return retList;
-      } else if (type == DType.STRING) {
-        int start = offsets.getInt(rowIndex * DType.INT32.getSizeInBytes());
-        int end = offsets.getInt((rowIndex + 1) * DType.INT32.getSizeInBytes());
-        int size = end - start;
-        byte[] rawData = new byte[size];
-        if (size > 0) {
-          data.getBytes(rawData, 0, start, size);
-          return new String(rawData);
-        } else if (isNull(rowIndex)) {
-          return null;
-        } else {
-          return new String();
-        }
-      } else {
-        if (isNull(rowIndex)) {
-          return null;
-        }
-        int start = rowIndex * type.getSizeInBytes();
-        return readValue(start);
-      }
-    }
-
-    public boolean isNull(long index) {
-      assert (index >= 0 && index < rows) : "index is out of range 0 <= " + index + " < " + rows;
-      if (validity != null) {
-        return BitVectorHelper.isNull(validity, index);
-      }
-      return false;
-    }
-
-    private Object readValue(int index){
-      assert index < rows * type.getSizeInBytes();
-      switch (type) {
-        case INT32: // fall through
-        case UINT32: // fall through
-        case TIMESTAMP_DAYS:
-        case DURATION_DAYS: return data.getInt(index);
-        case INT64: // fall through
-        case UINT64: // fall through
-        case DURATION_MICROSECONDS: // fall through
-        case DURATION_MILLISECONDS: // fall through
-        case DURATION_NANOSECONDS: // fall through
-        case DURATION_SECONDS: // fall through
-        case TIMESTAMP_MICROSECONDS: // fall through
-        case TIMESTAMP_MILLISECONDS: // fall through
-        case TIMESTAMP_NANOSECONDS: // fall through
-        case TIMESTAMP_SECONDS: return data.getLong(index);
-        case FLOAT32: return data.getFloat(index);
-        case FLOAT64: return data.getDouble(index);
-        case UINT8: // fall through
-        case INT8: return data.getByte(index);
-        case UINT16: // fall through
-        case INT16: return data.getShort(index);
-        case BOOL8: return data.getBoolean(index);
-        default: throw new UnsupportedOperationException("Do not support " + type);
-      }
-    }
-
-    public long getHostMemorySize() {
-      long totalSize = 0;
-      if (data != null) {
-        totalSize += data.length;
-      }
-      if (offsets != null) {
-        totalSize += offsets.length;
-      }
-      if (data != null) {
-        totalSize += validity.length;
-      }
-      for (NestedHostColumnVector nhcv : nestedChildren) {
-        totalSize += nhcv.getHostMemorySize();
-      }
-      return totalSize;
-    }
-  }
-
-  private final OffHeapState offHeap;
-  private final DType type;
-  private long rows;
-  private Optional<Long> nullCount = Optional.empty();
   private int refCount;
-  List<NestedHostColumnVector> children = new ArrayList<>();
 
   /**
    * Create a new column vector with data populated on the host.
@@ -225,19 +61,14 @@ public final class HostColumnVector implements AutoCloseable {
    * @param offsetBuffer       only valid for STRING this is the offsets into
    *                           the hostDataBuffer indicating the start and end of a string
    *                           entry. It should be (rows + 1) ints.
-   * @param nestedHcv          list of child NestedHostColumnVector(s) for complex types
+   * @param nestedHcv          list of child BaseHostColumnVector(s) for complex types
    */
 
   //Constructor for lists
   HostColumnVector(DType type, long rows, Optional<Long> nullCount,
                    HostMemoryBuffer hostDataBuffer, HostMemoryBuffer hostValidityBuffer,
-                   HostMemoryBuffer offsetBuffer, List<NestedHostColumnVector> nestedHcv) {
-    this.offHeap = new OffHeapState(hostDataBuffer, hostValidityBuffer, offsetBuffer);
-    MemoryCleaner.register(this, offHeap);
-    this.rows = rows;
-    this.nullCount = nullCount;
-    this.type = type;
-    this.children = nestedHcv;
+                   HostMemoryBuffer offsetBuffer, List<BaseHostColumnVector> nestedHcv) {
+    super(type, rows, nullCount, hostDataBuffer, hostValidityBuffer, offsetBuffer, nestedHcv);
     refCount = 0;
     incRefCountInternal(true);
   }
@@ -245,6 +76,7 @@ public final class HostColumnVector implements AutoCloseable {
   HostColumnVector(DType type, long rows, Optional<Long> nullCount,
                    HostMemoryBuffer hostDataBuffer, HostMemoryBuffer hostValidityBuffer,
                    HostMemoryBuffer offsetBuffer) {
+    super(type, rows, nullCount, hostDataBuffer, hostValidityBuffer, offsetBuffer, new ArrayList<>());
     assert type != DType.LIST : "This constructor should not be used for list type";
     if (nullCount.isPresent() && nullCount.get() > 0 && hostValidityBuffer == null) {
       throw new IllegalStateException("Buffer cannot have a nullCount without a validity buffer");
@@ -252,12 +84,6 @@ public final class HostColumnVector implements AutoCloseable {
     if (type != DType.STRING && type != DType.LIST) {
       assert offsetBuffer == null : "offsets are only supported for STRING and LIST";
     }
-    offHeap = new OffHeapState(hostDataBuffer, hostValidityBuffer, offsetBuffer);
-    MemoryCleaner.register(this, offHeap);
-    this.rows = rows;
-    this.nullCount = nullCount;
-    this.type = type;
-
     refCount = 0;
     incRefCountInternal(true);
   }
@@ -281,7 +107,7 @@ public final class HostColumnVector implements AutoCloseable {
     offHeap.delRef();
     if (refCount == 0) {
       offHeap.clean(false);
-      for( NestedHostColumnVector child : children) {
+      for( BaseHostColumnVector child : children) {
         child.close();
       }
     } else if (refCount < 0) {
@@ -335,7 +161,7 @@ public final class HostColumnVector implements AutoCloseable {
   public long getHostMemorySize() {
     long childTotalSize = 0;
     if (!children.isEmpty()) {
-      for (NestedHostColumnVector nhcv : children) {
+      for (BaseHostColumnVector nhcv : children) {
         childTotalSize += nhcv.getHostMemorySize();
       }
     }
@@ -667,90 +493,7 @@ public final class HostColumnVector implements AutoCloseable {
     return new String(rawData, StandardCharsets.UTF_8);
   }
 
-  /////////////////////////////////////////////////////////////////////////////
-  // HELPER CLASSES
-  /////////////////////////////////////////////////////////////////////////////
 
-  /**
-   * Holds the off heap state of the column vector so we can clean it up, even if it is leaked.
-   */
-  protected static final class OffHeapState extends MemoryCleaner.Cleaner {
-    public HostMemoryBuffer data;
-    public HostMemoryBuffer valid = null;
-    public HostMemoryBuffer offsets = null;
-
-    OffHeapState(HostMemoryBuffer data, HostMemoryBuffer valid, HostMemoryBuffer offsets) {
-      this.data = data;
-      this.valid = valid;
-      this.offsets = offsets;
-    }
-
-    @Override
-    protected boolean cleanImpl(boolean logErrorIfNotClean) {
-      boolean neededCleanup = false;
-      if (data != null || valid != null || offsets != null) {
-        try {
-          ColumnVector.closeBuffers(data);
-          ColumnVector.closeBuffers(offsets);
-          ColumnVector.closeBuffers(valid);
-        } finally {
-          // Always mark the resource as freed even if an exception is thrown.
-          // We cannot know how far it progressed before the exception, and
-          // therefore it is unsafe to retry.
-          data = null;
-          valid = null;
-          offsets = null;
-        }
-        neededCleanup = true;
-      }
-      if (neededCleanup && logErrorIfNotClean) {
-        log.error("A HOST COLUMN VECTOR WAS LEAKED (ID: " + id + ")");
-        logRefCountDebug("Leaked vector");
-      }
-      return neededCleanup;
-    }
-
-    @Override
-    public void noWarnLeakExpected() {
-      super.noWarnLeakExpected();
-      if (data != null) {
-        data.noWarnLeakExpected();
-      }
-      if (valid != null) {
-        valid.noWarnLeakExpected();
-      }
-      if (offsets != null) {
-        offsets.noWarnLeakExpected();
-      }
-    }
-
-    @Override
-    public boolean isClean() {
-      return data == null && valid == null && offsets == null;
-    }
-
-    /**
-     * This returns total memory allocated on the host for the ColumnVector.
-     */
-    public long getHostMemorySize() {
-      long total = 0;
-      if (valid != null) {
-        total += valid.length;
-      }
-      if (data != null) {
-        total += data.length;
-      }
-      if (offsets != null) {
-        total += offsets.length;
-      }
-      return total;
-    }
-
-    @Override
-    public String toString() {
-      return "(ID: " + id + ")";
-    }
-  }
 
   /////////////////////////////////////////////////////////////////////////////
   // BUILDER
@@ -1244,21 +987,21 @@ public final class HostColumnVector implements AutoCloseable {
     }
 
     public HostColumnVector build() {
-      List<NestedHostColumnVector> nestedHostColumnVectorList = new ArrayList<>();
+      List<BaseHostColumnVector> BaseHostColumnVectorList = new ArrayList<>();
       for (ColumnBuilder childBuilder : childBuilders) {
-        nestedHostColumnVectorList.add(childBuilder.buildNestedInternal());
+        BaseHostColumnVectorList.add(childBuilder.buildNestedInternal());
       }
       HostColumnVector hostColumnVector = new HostColumnVector(type, rows, Optional.of(nullCount), data, valid, offsets,
-          nestedHostColumnVectorList);
+          BaseHostColumnVectorList);
       return hostColumnVector;
     }
 
-    private NestedHostColumnVector buildNestedInternal() {
-      List<NestedHostColumnVector> nestedHostColumnVectorList = new ArrayList<>();
+    private BaseHostColumnVector buildNestedInternal() {
+      List<BaseHostColumnVector> baseHostColumnVectorList = new ArrayList<>();
       for (ColumnBuilder childBuilder : childBuilders) {
-        nestedHostColumnVectorList.add(childBuilder.buildNestedInternal());
+        baseHostColumnVectorList.add(childBuilder.buildNestedInternal());
       }
-      NestedHostColumnVector ret = new NestedHostColumnVector(type, rows, Optional.of(nullCount), data, valid, offsets, nestedHostColumnVectorList);
+      BaseHostColumnVector ret = new BaseHostColumnVector(type, rows, Optional.of(nullCount), data, valid, offsets, baseHostColumnVectorList);
       return ret;
     }
