+++ java/src/main/java/ai/rapids/cudf/MemoryCleaner.java
@@ -219,7 +219,7 @@ final class MemoryCleaner {
     all.add(new CleanerWeakReference(vec, cleaner, collected, true));
   }
 
-  static void register(HostColumnVector vec, Cleaner cleaner) {
+  static void register(BaseHostColumnVector vec, Cleaner cleaner) {
     // It is now registered...
     all.add(new CleanerWeakReference(vec, cleaner, collected, false));
   }
